Move duplicated v0/v1 impl packages to v0 subpackages
Fixes #2616. The convention is adopted from the command API (v1)
whose impl classes have the v1 subpackage.
