+++ deprecated/fabric-events-lifecycle-v0/src/client/java/net/fabricmc/fabric/impl/event/lifecycle/v0/client/LegacyClientEventInvokers.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.event.lifecycle.client;
+package net.fabricmc.fabric.impl.event.lifecycle.v0.client;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
