+++ deprecated/fabric-networking-v0/src/client/java/net/fabricmc/fabric/impl/networking/v0/ClientSidePacketRegistryImpl.java
@@ -14,72 +14,69 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.networking;
+package net.fabricmc.fabric.impl.networking.v0;
 
 import java.util.Objects;
 
 import io.netty.util.concurrent.Future;
 import io.netty.util.concurrent.GenericFutureListener;
 
+import net.fabricmc.fabric.impl.networking.GenericFutureListenerHolder;
+import net.minecraft.client.MinecraftClient;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.Packet;
 import net.minecraft.network.PacketByteBuf;
-import net.minecraft.network.packet.s2c.play.CustomPayloadS2CPacket;
-import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.thread.ThreadExecutor;
 
 import net.fabricmc.api.EnvType;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
+import net.fabricmc.fabric.api.network.ClientSidePacketRegistry;
 import net.fabricmc.fabric.api.network.PacketConsumer;
 import net.fabricmc.fabric.api.network.PacketContext;
 import net.fabricmc.fabric.api.network.PacketRegistry;
-import net.fabricmc.fabric.api.network.ServerSidePacketRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 
-public class ServerSidePacketRegistryImpl implements ServerSidePacketRegistry, PacketRegistry {
+public class ClientSidePacketRegistryImpl implements ClientSidePacketRegistry, PacketRegistry {
 	@Override
-	public boolean canPlayerReceive(PlayerEntity player, Identifier id) {
-		if (player instanceof ServerPlayerEntity) {
-			return ServerPlayNetworking.canSend((ServerPlayerEntity) player, id);
-		}
-
-		return false;
+	public boolean canServerReceive(Identifier id) {
+		return ClientPlayNetworking.getSendable().contains(id);
 	}
 
 	@Override
-	public void sendToPlayer(PlayerEntity player, Packet<?> packet, GenericFutureListener<? extends Future<? super Void>> completionListener) {
-		if (player instanceof ServerPlayerEntity) {
-			((ServerPlayerEntity) player).networkHandler.sendPacket(packet, GenericFutureListenerHolder.create(completionListener));
+	public void sendToServer(Packet<?> packet, GenericFutureListener<? extends Future<? super Void>> completionListener) {
+		if (MinecraftClient.getInstance().getNetworkHandler() != null) {
+			MinecraftClient.getInstance().getNetworkHandler().getConnection().send(packet, GenericFutureListenerHolder.create(completionListener));
 			return;
 		}
 
-		throw new RuntimeException("Can only send to ServerPlayerEntities!");
+		throw new IllegalStateException("Cannot send packet to server while not in game!"); // TODO: Error message
 	}
 
 	@Override
 	public Packet<?> toPacket(Identifier id, PacketByteBuf buf) {
-		return new CustomPayloadS2CPacket(id, buf);
+		return ClientPlayNetworking.createC2SPacket(id, buf);
 	}
 
 	@Override
 	public void register(Identifier id, PacketConsumer consumer) {
+		// id is checked in client networking
 		Objects.requireNonNull(consumer, "PacketConsumer cannot be null");
 
-		ServerPlayNetworking.registerGlobalReceiver(id, (server, player, handler, buf, sender) -> {
+		ClientPlayNetworking.registerGlobalReceiver(id, (client, handler, buf, sender) -> {
 			consumer.accept(new PacketContext() {
 				@Override
 				public EnvType getPacketEnvironment() {
-					return EnvType.SERVER;
+					return EnvType.CLIENT;
 				}
 
 				@Override
 				public PlayerEntity getPlayer() {
-					return player;
+					return client.player;
 				}
 
 				@Override
 				public ThreadExecutor<?> getTaskQueue() {
-					return server;
+					return client;
 				}
 			}, buf);
 		});
@@ -87,6 +84,6 @@ public class ServerSidePacketRegistryImpl implements ServerSidePacketRegistry, P
 
 	@Override
 	public void unregister(Identifier id) {
-		ServerPlayNetworking.unregisterGlobalReceiver(id);
+		ClientPlayNetworking.unregisterGlobalReceiver(id);
 	}
 }
